feat(order, user): adapting route to delete user and order

+++ src/services/order/DeleteOrderService.ts
@@ -1,3 +1,4 @@
+import createHttpError from "http-errors";
 import { injectable } from "tsyringe";
 import { OrderRepository } from "../../repositories/OrderRepository";
 
@@ -7,7 +8,13 @@ export class DeleteOrderService {
     private orderRepository: OrderRepository,
   ) {}
 
-  async execute(id: string) {
+  async execute(id: string, userId: string) {
+    const found = await this.orderRepository.find(id, userId);
+
+    if(!found){
+      throw new createHttpError.NotFound("Order Not Found");
+    }
+    
     return this.orderRepository.delete(id)
   }
 }
